+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # CHANGELOG
 
-NEXT_VERSION_BUMP: PATCH
+NEXT_VERSION_BUMP: MINOR
 ## XX.XX.XX - 20XX-XX-XX
 
 ### PaymentSheet
 * [FIXED][14436](https://github.com/stripe/stripe-android/pull/14436) Fixed a rare race condition which could result in the payment sheet being half-visible on screen.
+
+### CryptoOnramp
+* [ADDED] Added `presentTermsAndConditionsIfNeeded()` to present and record terms acceptance only when required.
+* [ADDED] Added `presentTermsOfServiceIfNeeded()` to present and record terms of service acceptance during onboarding only when required.
 
 ## 23.18.0 - 2026-09-08
 
@@ -20,8 +24,6 @@ NEXT_VERSION_BUMP: PATCH
 * [CHANGED] Use Stripe-hosted address autocomplete by default.
 
 ### CryptoOnramp
-* [ADDED] Added `presentTermsAndConditionsIfNeeded()` to present and record terms acceptance only when required.
-* [ADDED] Added `presentTermsOfServiceIfNeeded()` to present and record terms of service acceptance during onboarding only when required.
 * [ADDED] Added Canada SIN, Colombia NIT, and Philippines TIN values to `IdType`, and added `idType` to `KycInfo`.
 
 ## 23.17.1 - 2026-08-31